+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased][]
+
+## [1.4.0][] - 2018-04-22
 - Allowed the user to set the title using -T
 - Allowed the user to set the tag using -t
 - Allowed the user to default to exporting the next month
@@ -33,7 +35,8 @@
 ### Added
 - Changelog versioning
 
-[Unreleased]: https://github.com/one19/monthTodo/compare/v1.3.0...HEAD
+[Unreleased]: https://github.com/one19/monthTodo/compare/v1.4.0...HEAD
+[1.4.0]: https://github.com/one19/monthTodo/compare/v1.3.0...v1.4.0
 [1.3.0]: https://github.com/one19/monthTodo/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/one19/monthTodo/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/one19/monthTodo/compare/v1.1.0...v1.1.1